+++ src/main/java/net/silentchaos512/endertendril/client/EnderTendrilClient.java
@@ -1,0 +1,11 @@
+package net.silentchaos512.endertendril.client;
+
+import net.neoforged.fml.ModContainer;
+import net.neoforged.neoforge.client.gui.ConfigurationScreen;
+import net.neoforged.neoforge.client.gui.IConfigScreenFactory;
+
+public class EnderTendrilClient {
+    public EnderTendrilClient(ModContainer modContainer) {
+        modContainer.registerExtensionPoint(IConfigScreenFactory.class, ConfigurationScreen::new);
+    }
+}