+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.21.1-1.7.0] - 2024-10-19
+- Ported to NeoForge 1.21.1 (supports 1.21 too)
+- Added a config option to change the growth speed of ender tendrils
+- Added a configuration screen (in the Mods menu)
+
 ## [1.20.1-1.6.1] - 2023-06-16
 - Updated to Minecraft 1.20.1